Docker container removed and created -- Final commit

+++ Jenkinsfile
@@ -115,26 +115,18 @@
             }
           }
 		
-		stage('Docker -- Stop & Removing Running Container') {
+        stage('Docker -- Stop & Removing Running Container') {
           steps{
-					bat "docker stop ProductManagementApi"
+			script {
+                def containerId = powershell(returnStdout: true, script: "docker ps -f name=ProductManagementApi   | Select-String 5000 | %{ (\$_ -split \" \")[0]}");
+                if(containerId!= null && containerId!="") {
+                    bat "docker stop ProductManagementApi"
 					bat "docker rm -f ProductManagementApi"
+                }
+            }
           }
-        }  		
-		//stage('Docker -- Stop Running Container') {
-          //steps{
-				//	bat """
-					//	 ContainerId = docker inspect --format="{{.Id}}" ProductManagementApi
-					//	echo $ContainerId
-						//if [ $ContainerId ]
-						//then 
-							//docker stop ${ContainerId}
-							//docker rm -f ${ContainerId}
-						//fi
-				//	"""
-           // }
-          //}  
-		  
+		}		  
+	  
 		stage('Docker Deployment') {
           steps{
                     bat "docker run --name ProductManagementApi -d -p 5000:80 ${registry}:${BUILD_NUMBER}"
